epos_repro_cluster
This file intends to optimize the day that need to be reprocessed for any of the following reason: High residual , ambiguities exceeded,cluster file or another reason user defined. Besides, the CLN_GLOBAL file that user define for "repro" propose will be updated based on the result of epos_repro_cluster. Eventually, the dates and error types selected by the user will be uploaded to the Section 1.1 GFZ cluster